feat/table: address code smells
