+++ client/src/app/api/ontimeApi.js
@@ -105,7 +105,7 @@ export const getSettings = async () => {
 };
 
 export const postSettings = async (data) => {
-  return await axios.post(`${ontimeURL}/settings`, data);
+  await axios.post(`${ontimeURL}/settings`, data);
 };
 
 export const getInfo = async () => {
@@ -114,7 +114,7 @@ export const getInfo = async () => {
 };
 
 export const postInfo = async (data) => {
-  return await axios.post(`${ontimeURL}/info`, data);
+  await axios.post(`${ontimeURL}/info`, data);
 };
 
 export const getAliases = async () => {
@@ -123,7 +123,7 @@ export const getAliases = async () => {
 };
 
 export const postAliases = async (data) => {
-  return await axios.post(`${ontimeURL}/aliases`, data);
+  await axios.post(`${ontimeURL}/aliases`, data);
 };
 
 
@@ -133,7 +133,7 @@ export const getUserFields = async () => {
 };
 
 export const postUserFields = async (data) => {
-  return await axios.post(`${ontimeURL}/userfields`, data);
+  await axios.post(`${ontimeURL}/userfields`, data);
 };
 
 export const getOSC = async () => {
@@ -142,7 +142,7 @@ export const getOSC = async () => {
 };
 
 export const postOSC = async (data) => {
-  return await axios.post(`${ontimeURL}/osc`, data);
+  await axios.post(`${ontimeURL}/osc`, data);
 };
 
 export const downloadEvents = async () => {
